Consolidated migrations to main migration
