--- lib/Migration/Version0010Date20191218203500.php
@@ -1,110 +1,0 @@
-<?php
-/**
- * @copyright Copyright (c) 2017 René Gieling <github@dartcafe.de>
- *
- * @author René Gieling <github@dartcafe.de>
- *
- * @license GNU AGPL version 3 or any later version
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-namespace OCA\Polls\Migration;
-
-// use Doctrine\DBAL\Exception\TableNotFoundException;
-// use Doctrine\DBAL\Platforms\PostgreSqlPlatform;
-use Doctrine\DBAL\Types\Type;
-use OCP\DB\ISchemaWrapper;
-use OCP\DB\QueryBuilder\IQueryBuilder;
-use OCP\IConfig;
-use OCP\IDBConnection;
-use OCP\Migration\SimpleMigrationStep;
-use OCP\Migration\IOutput;
-use OCP\Security\ISecureRandom;
-
-/**
- * Installation class for the polls app.
- * Initial db creation
- */
-class Version0010Date20191218203500 extends SimpleMigrationStep {
-
-	/** @var IDBConnection */
-	protected $connection;
-
-	/** @var IConfig */
-	protected $config;
-
-	/**
-	 * @param IDBConnection $connection
-	 * @param IConfig $config
-	 */
-	public function __construct(IDBConnection $connection, IConfig $config) {
-		$this->connection = $connection;
-		$this->config = $config;
-	}
-
-	/**
-	 * @param IOutput $output
-	 * @param \Closure $schemaClosure The `\Closure` returns a `ISchemaWrapper`
-	 * @param array $options
-	 * @return null|ISchemaWrapper
-	 * @since 13.0.0
-	 */
-	public function changeSchema(IOutput $output, \Closure $schemaClosure, array $options) {
-		/** @var ISchemaWrapper $schema */
-		$schema = $schemaClosure();
-
-		if ($schema->hasTable('polls_events')) {
-			$table = $schema->getTable('polls_events');
-			if (!$table->hasColumn('expiration')) {
-				$table->addColumn('expiration', Type::BOOLEAN, [
-					'notnull' => true,
-					'default' => 0
-				]);
-			}
-
-		}
-
-		return $schema;
-	}
-
-	/**
-	 * @param IOutput $output
-	 * @param \Closure $schemaClosure The `\Closure` returns a `ISchemaWrapper`
-	 * @param array $options
-	 * @since 13.0.0
-	 */
-	public function postSchemaChange(IOutput $output, \Closure $schemaClosure, array $options) {
-		/** @var ISchemaWrapper $schema */
-		$schema = $schemaClosure();
-
-		if ($schema->hasTable('polls_events')) {
-			$this->setExpiration();
-		}
-	}
-
-	/**
-	 * Copy public tokens
-	 */
-	protected function setExpiration() {
-
-		$update = $this->connection->getQueryBuilder();
-		$update->update('polls_events')
-			->set('expiration', $update->createNamedParameter(true))
-			->where('expire IS NOT NULL');
-		$result = $update->execute();
-	}
-
-}
